Splitting out Lua API functions
