+++ TIBERIANDAWN/LUA_API_FUNCTIONS_AIR_RULES.H
@@ -1,0 +1,6 @@
+#ifndef LUA_API_FUNCTIONS_AIR_RULES_H
+#define LUA_API_FUNCTIONS_AIR_RULES_H
+
+bool Register_Aircraft_Rule_Functions();
+
+#endif
