+++ TIBERIANDAWN/LUA_API_FUNCTIONS_GAME.H
@@ -1,0 +1,6 @@
+#ifndef LUA_API_FUNCTIONS_H
+#define LUA_API_FUNCTIONS_H
+
+bool Register_Game_Functions();
+
+#endif
