+++ TIBERIANDAWN/LUA_API_FUNCTIONS_GAME_RULES.H
@@ -1,0 +1,6 @@
+#ifndef LUA_API_FUNCTIONS_GAME_RULES_H
+#define LUA_API_FUNCTIONS_GAME_RULES_H
+
+bool Register_Game_Rule_Functions();
+
+#endif
