+++ TIBERIANDAWN/LUA_API_FUNCTIONS_INFO.H
@@ -1,0 +1,6 @@
+#ifndef LUA_API_FUNCTIONS_INFO_H
+#define LUA_API_FUNCTIONS_INFO_H
+
+bool Register_Info_Functions();
+
+#endif
