+++ TIBERIANDAWN/LUA_API_FUNCTIONS_INF_RULES.H
@@ -1,0 +1,6 @@
+#ifndef LUA_API_FUNCTIONS_INF_RULES_H
+#define LUA_API_FUNCTIONS_INF_RULES_H
+
+bool Register_Infantry_Rule_Functions();
+
+#endif
