+++ TIBERIANDAWN/LUA_API_FUNCTIONS_UTIL.CPP
@@ -1,0 +1,83 @@
+#include "logger.h"
+#include "lua.h"
+#include "rules_ini.h"
+#include "strings.h"
+#include "utils.h"
+
+static int Lua_Get_Now_In_Epoch_Millis(lua_State* lua)
+{
+    lua_pushnumber(lua, Get_Now_In_Epoch_Millis());
+
+    return 1;
+}
+
+static int Lua_Set_Log_Level(lua_State* lua)
+{
+    int argCount = lua_gettop(lua);
+
+    if (argCount > 0)
+    {
+        auto logLevelStr = Get_Lua_Value_As_String(1);
+
+        auto logLevelUpper = Convert_String_To_Upper_Case(logLevelStr);
+
+        auto logLevel = Parse_Log_Level(logLevelUpper);
+
+        delete logLevelUpper;
+
+        Set_Current_Log_Level(logLevel);
+    }
+
+    return 0;
+}
+
+static int Lua_Get_Log_Level(lua_State* lua)
+{
+    auto logLevel = Log_Level_To_String(Current_Log_Level());
+
+    lua_pushstring(lua, logLevel);
+
+    return 1;
+}
+
+static int Lua_Show_Error(lua_State* lua)
+{
+    int argCount = lua_gettop(lua);
+
+    if (argCount > 0)
+    {
+        auto logMessage = Get_Lua_Value_As_String(1);
+
+        Show_Error(logMessage);
+    }
+
+    return 0;
+}
+
+static int Lua_Log(lua_State* lua)
+{
+    int argCount = lua_gettop(lua);
+
+    if (argCount > 0)
+    {
+        auto logMessage = Get_Lua_Value_As_String(1);
+
+        Log_Info("Lua => %s", logMessage);
+    }
+
+    return 0;
+}
+
+bool Register_Util_Functions()
+{
+    Log_Debug("Initialising Lua API utility functions");
+
+    Register_Lua_Function("log", Lua_Log);
+    Register_Lua_Function("print", Lua_Log);
+    Register_Lua_Function("showError", Lua_Show_Error);
+    Register_Lua_Function("getLogLevel", Lua_Get_Log_Level);
+    Register_Lua_Function("setLogLevel", Lua_Set_Log_Level);
+    Register_Lua_Function("getNowInEpochMillis", Lua_Get_Now_In_Epoch_Millis);
+
+    return true;
+}
