+++ TIBERIANDAWN/LUA_API_FUNCTIONS_UTIL.H
@@ -1,0 +1,6 @@
+#ifndef LUA_API_FUNCTIONS_UTILS_H
+#define LUA_API_FUNCTIONS_UTILS_H
+
+bool Register_Util_Functions();
+
+#endif
