+++ TIBERIANDAWN/NCO.H
@@ -4,8 +4,6 @@
 #include "./logger.h"
 #include "./lua.h"
 #include "./lua_events.h"
-#include "./lua_api.h"
-#include "./lua_api_functions.h"
 #include "./nco.h"
 #include "./parse.h"
 #include "./rules_ini.h"
