+++ TIBERIANDAWN/PARSE.cpp
@@ -637,118 +637,6 @@ char* Weapon_Type_To_String(WeaponType weaponType)
     return weaponTypeString;
 }
 
-char* Bullet_Type_To_String(BulletType type)
-{
-    char* bulletTypeString;
-
-    if (type == BULLET_SNIPER)
-    {
-        bulletTypeString = "SNIPER";
-    }
-    else if (type == BULLET_SPREADFIRE)
-    {
-        bulletTypeString = "SPREADFIRE";
-    }
-    else if (type == BULLET_BULLET)
-    {
-        bulletTypeString = "BULLET";
-    }
-    else if (type == BULLET_BULLET)
-    {
-        bulletTypeString = "BULLET";
-    }
-    else if (type == BULLET_TOW)
-    {
-        bulletTypeString = "TOW";
-    }
-    else if (type == BULLET_FLAME)
-    {
-        bulletTypeString = "FLAME";
-    }
-    else if (type == BULLET_FLAME)
-    {
-        bulletTypeString = "FLAME";
-    }
-    else if (type == BULLET_CHEMSPRAY)
-    {
-        bulletTypeString = "CHEMSPRAY";
-    }
-    else if (type == BULLET_GRENADE)
-    {
-        bulletTypeString = "GRENADE";
-    }
-    else if (type == BULLET_APDS)
-    {
-        bulletTypeString = "APDS";
-    }
-    else if (type == BULLET_APDS)
-    {
-        bulletTypeString = "APDS";
-    }
-    else if (type == BULLET_APDS)
-    {
-        bulletTypeString = "APDS";
-    }
-    else if (type == BULLET_APDS)
-    {
-        bulletTypeString = "APDS";
-    }
-    else if (type == BULLET_SSM)
-    {
-        bulletTypeString = "SSM";
-    }
-    else if (type == BULLET_SSM2)
-    {
-        bulletTypeString = "SSM2";
-    }
-    else if (type == BULLET_HE)
-    {
-        bulletTypeString = "HE";
-    }
-    else if (type == BULLET_BULLET)
-    {
-        bulletTypeString = "BULLET";
-    }
-    else if (type == BULLET_SSM)
-    {
-        bulletTypeString = "SSM";
-    }
-    else if (type == BULLET_SSM)
-    {
-        bulletTypeString = "SSM";
-    }
-    else if (type == BULLET_NAPALM)
-    {
-        bulletTypeString = "NAPALM";
-    }
-    else if (type == BULLET_LASER)
-    {
-        bulletTypeString = "LASER";
-    }
-    else if (type == BULLET_SAM)
-    {
-        bulletTypeString = "SAM";
-    }
-    else if (type == BULLET_HONEST_JOHN)
-    {
-        bulletTypeString = "HONEST_JOHN";
-    }
-    else if (type == BULLET_HEADBUTT)
-    {
-        bulletTypeString = "HEADBUTT";
-    }
-    else if (type == BULLET_TREXBITE)
-    {
-        bulletTypeString = "TREXBITE";
-    }
-    else
-    {
-        Show_Error("Unable to convert bullet type to string: %d", type);
-    }
-
-    return bulletTypeString;
-}
-
 ArmorType Parse_Armor_Type(char* armorTypeString, bool* parseError)
 {
     ArmorType armor;
