+++ TIBERIANDAWN/TiberianDawn.vcxproj
@@ -397,9 +397,17 @@
     <ClInclude Include="LOGIC.H" />
     <ClInclude Include="LUA.H" />
     <ClInclude Include="LUA_API.H" />
-    <ClInclude Include="LUA_API_FUNCTIONS.H" />
+    <ClInclude Include="LUA_API_FUNCTIONS_GAME.H" />
+    <ClInclude Include="LUA_API_FUNCTIONS_AIR_RULES.H" />
+    <ClInclude Include="LUA_API_FUNCTIONS_BUILD_RULES.H" />
+    <ClInclude Include="LUA_API_FUNCTIONS_GAME_RULES.H" />
+    <ClInclude Include="LUA_API_FUNCTIONS_GENERIC_RULES.H" />
+    <ClInclude Include="LUA_API_FUNCTIONS_INF_RULES.H" />
+    <ClInclude Include="LUA_API_FUNCTIONS_UNIT_RULES.H" />
+    <ClInclude Include="LUA_API_FUNCTIONS_UTIL.H" />
     <ClInclude Include="LUA_EVENTS.H" />
     <ClInclude Include="LUA_REPL.H" />
+    <ClInclude Include="LUA_API_FUNCTIONS_INFO.H" />
     <ClInclude Include="MAP.H" />
     <ClInclude Include="MAPEDIT.H" />
     <ClInclude Include="MEMCHECK.H" />
@@ -580,7 +588,15 @@
     <ClCompile Include="EDIT.CPP" />
     <ClCompile Include="ENDING.CPP" />
     <ClCompile Include="LUA_API.CPP" />
-    <ClCompile Include="LUA_API_FUNCTIONS.CPP" />
+    <ClCompile Include="LUA_API_FUNCTIONS_GAME.CPP" />
+    <ClCompile Include="LUA_API_FUNCTIONS_AIR_RULES.CPP" />
+    <ClCompile Include="LUA_API_FUNCTIONS_BUILD_RULES.CPP" />
+    <ClCompile Include="LUA_API_FUNCTIONS_INFO.CPP" />
+    <ClCompile Include="LUA_API_FUNCTIONS_GAME_RULES.CPP" />
+    <ClCompile Include="LUA_API_FUNCTIONS_GENERIC_RULES.CPP" />
+    <ClCompile Include="LUA_API_FUNCTIONS_INF_RULES.CPP" />
+    <ClCompile Include="LUA_API_FUNCTIONS_UNIT_RULES.CPP" />
+    <ClCompile Include="LUA_API_FUNCTIONS_UTIL.CPP" />
     <ClCompile Include="LUA_REPL.CPP" />
     <ClCompile Include="NCO.cpp" />
     <ClCompile Include="RULES_CACHE.CPP" />
