+++ TIBERIANDAWN/TiberianDawn.vcxproj.filters
@@ -700,9 +700,6 @@
     <ClInclude Include="NCO.H">
       <Filter>Source Files\NCO</Filter>
     </ClInclude>
-    <ClInclude Include="LUA_API_FUNCTIONS.H">
-      <Filter>Source Files\NCO</Filter>
-    </ClInclude>
     <ClInclude Include="LUA_REPL.H">
       <Filter>Source Files\NCO</Filter>
     </ClInclude>
@@ -734,6 +731,33 @@
       <Filter>Source Files\NCO</Filter>
     </ClInclude>
     <ClInclude Include="RULES_INI_SUPERWEAPON.H">
+      <Filter>Source Files\NCO</Filter>
+    </ClInclude>
+    <ClInclude Include="LUA_API_FUNCTIONS_INFO.H">
+      <Filter>Source Files\NCO</Filter>
+    </ClInclude>
+    <ClInclude Include="LUA_API_FUNCTIONS_GAME_RULES.H">
+      <Filter>Source Files\NCO</Filter>
+    </ClInclude>
+    <ClInclude Include="LUA_API_FUNCTIONS_GENERIC_RULES.H">
+      <Filter>Source Files\NCO</Filter>
+    </ClInclude>
+    <ClInclude Include="LUA_API_FUNCTIONS_INF_RULES.H">
+      <Filter>Source Files\NCO</Filter>
+    </ClInclude>
+    <ClInclude Include="LUA_API_FUNCTIONS_UNIT_RULES.H">
+      <Filter>Source Files\NCO</Filter>
+    </ClInclude>
+    <ClInclude Include="LUA_API_FUNCTIONS_AIR_RULES.H">
+      <Filter>Source Files\NCO</Filter>
+    </ClInclude>
+    <ClInclude Include="LUA_API_FUNCTIONS_BUILD_RULES.H">
+      <Filter>Source Files\NCO</Filter>
+    </ClInclude>
+    <ClInclude Include="LUA_API_FUNCTIONS_UTIL.H">
+      <Filter>Source Files\NCO</Filter>
+    </ClInclude>
+    <ClInclude Include="LUA_API_FUNCTIONS_GAME.H">
       <Filter>Source Files\NCO</Filter>
     </ClInclude>
   </ItemGroup>
@@ -1345,13 +1369,37 @@
     <ClCompile Include="NCO.cpp">
       <Filter>Source Files\NCO</Filter>
     </ClCompile>
-    <ClCompile Include="LUA_API_FUNCTIONS.CPP">
-      <Filter>Source Files\NCO</Filter>
-    </ClCompile>
     <ClCompile Include="LUA_REPL.CPP">
       <Filter>Source Files\NCO</Filter>
     </ClCompile>
     <ClCompile Include="RULES_CACHE.CPP">
+      <Filter>Source Files\NCO</Filter>
+    </ClCompile>
+    <ClCompile Include="LUA_API_FUNCTIONS_INFO.CPP">
+      <Filter>Source Files\NCO</Filter>
+    </ClCompile>
+    <ClCompile Include="LUA_API_FUNCTIONS_GAME_RULES.CPP">
+      <Filter>Source Files\NCO</Filter>
+    </ClCompile>
+    <ClCompile Include="LUA_API_FUNCTIONS_GENERIC_RULES.CPP">
+      <Filter>Source Files\NCO</Filter>
+    </ClCompile>
+    <ClCompile Include="LUA_API_FUNCTIONS_INF_RULES.CPP">
+      <Filter>Source Files\NCO</Filter>
+    </ClCompile>
+    <ClCompile Include="LUA_API_FUNCTIONS_UNIT_RULES.CPP">
+      <Filter>Source Files\NCO</Filter>
+    </ClCompile>
+    <ClCompile Include="LUA_API_FUNCTIONS_AIR_RULES.CPP">
+      <Filter>Source Files\NCO</Filter>
+    </ClCompile>
+    <ClCompile Include="LUA_API_FUNCTIONS_BUILD_RULES.CPP">
+      <Filter>Source Files\NCO</Filter>
+    </ClCompile>
+    <ClCompile Include="LUA_API_FUNCTIONS_UTIL.CPP">
+      <Filter>Source Files\NCO</Filter>
+    </ClCompile>
+    <ClCompile Include="LUA_API_FUNCTIONS_GAME.CPP">
       <Filter>Source Files\NCO</Filter>
     </ClCompile>
   </ItemGroup>
